requirements.txt
fastapi==0.103.1
uvicorn==0.22.0
python-multipart==0.0.5
numpy==1.21.6
librosa==0.9.2
tensorflow-cpu==2.8.0
shutil==latest